+++ entry/tests/test_model.py
@@ -1583,6 +1583,11 @@ class ModelTest(AironeTestCase):
         self.assertEqual(ret['ret_count'], 1)
         self.assertEqual(ret['ret_values'][0]['entry']['name'], 'e-10')
 
+        # check functionallity of the 'hint_attr_value' parameter
+        ret = Entry.search_entries(user, [entity.id], hint_attr_value='foo-0')
+        self.assertEqual(ret['ret_count'], 1)
+        self.assertEqual(ret['ret_values'][0]['entry']['name'], 'e-0')
+
     def test_search_entries_with_hint_referral(self):
         user = User.objects.create(username='hoge')