+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## In development
 ### Added
 * Added `^` and `$` operators on filtering `text` and `string` typed attributes in advanced search result
+* Fixed a search query timeout for long keywords
 
 ## v3.0.0